Cleaning up merge and finishing renaming
