+++ tests/behaviour/steps/inspectcli.py
@@ -9,7 +9,7 @@ from csvcubed.cli.inspectcsvw.metadataprinter import MetadataPrinter
 from csvcubed.utils.sparql_handler.code_list_repository import CodeListRepository
 from csvcubed.utils.sparql_handler.csvw_repository import CsvWRepository
 from csvcubed.utils.sparql_handler.data_cube_repository import DataCubeRepository
-from tests.helpers.inspectors_cache import get_csvw_rdf_manager
+from tests.helpers.repository_cache import get_csvw_rdf_manager
 
 
 def _unformat_multiline_string(string: str) -> str:
