+++ tests/test-cases/cli/inspect/repository-load-dataframe/pivoted-shape/pivoted-shape-out/testing-converting-a-pivoted-csvw-to-pandas-dataframe.csv-metadata.json
