+++ tests/test-cases/cli/inspect/repository-load-dataframe/pivoted-shape/pivoted-shape.json
