+++ tests/test-cases/cli/inspect/repository-load-dataframe/standard-shape/standard-shape-out/dim2.csv
