+++ tests/test-cases/cli/inspect/repository-load-dataframe/standard-shape/standard-shape-out/testing-converting-a-standard-shape-csvw-to-pandas-dataframe.csv-metadata.json
