+++ tests/unit/cli/inspectcsvw/test_metadatainputvalidator.py
@@ -5,7 +5,7 @@ import pytest
 
 from csvcubed.cli.inspectcsvw.metadatainputvalidator import MetadataValidator
 from csvcubed.models.csvwtype import CSVWType
-from tests.helpers.inspectors_cache import get_csvw_rdf_manager
+from tests.helpers.repository_cache import get_csvw_rdf_manager
 from tests.unit.test_baseunit import get_test_cases_dir
 
 _test_case_base_dir = get_test_cases_dir() / "cli" / "inspect"
