+++ tests/unit/utils/sparqlhandler/test_code_list_repository.py
@@ -1,7 +1,10 @@
 import pytest
 
 from csvcubed.models.sparqlresults import CatalogMetadataResult, CodeListTableIdentifers
-from tests.helpers.inspectors_cache import get_code_list_inspector, get_csvw_rdf_manager
+from tests.helpers.repository_cache import (
+    get_code_list_repository,
+    get_csvw_rdf_manager,
+)
 from tests.unit.test_baseunit import get_test_cases_dir
 
 _test_case_base_dir = get_test_cases_dir() / "cli" / "inspect"
@@ -16,8 +19,8 @@ def test_code_list_table_identifiers():
         / "pivoted-multi-measure-dataset"
         / "qb-id-10003.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
-    code_list_identifiers = code_list_inspector._code_list_table_identifiers
+    code_list_repository = get_code_list_repository(path_to_cube)
+    code_list_identifiers = code_list_repository._code_list_table_identifiers
 
     assert len(code_list_identifiers) == 1
     assert code_list_identifiers[0] == CodeListTableIdentifers(
@@ -32,10 +35,10 @@ def test_code_list_table_identifiers_error():
     path_to_cube = (
         _test_case_base_dir / "itis-industry-multiple-skos-inscheme.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     with pytest.raises(KeyError) as exception:
-        _ = code_list_inspector._code_list_table_identifiers
+        _ = code_list_repository._code_list_table_identifiers
 
     assert (
         "Found multiple skos:inScheme columns in 'itis-industry-multiple-skos-inscheme.csv'."
@@ -52,9 +55,9 @@ def test_get_catalog_metadata_for_concept_scheme():
         / "pivoted-multi-measure-dataset"
         / "qb-id-10003.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
-    the_result = code_list_inspector.get_catalog_metadata_for_concept_scheme(
+    the_result = code_list_repository.get_catalog_metadata_for_concept_scheme(
         "some-dimension.csv#code-list"
     )
 
@@ -69,10 +72,10 @@ def test_get_catalog_metadata_for_concept_scheme_error():
     path_to_cube = (
         _test_case_base_dir / "itis-industry-no-skos-inscheme.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     with pytest.raises(KeyError) as exception:
-        _ = code_list_inspector.get_catalog_metadata_for_concept_scheme(
+        _ = code_list_repository.get_catalog_metadata_for_concept_scheme(
             "http://gss-data.org.uk/data/gss_data/trade/ons-international-trade-in-services#scheme/it-industry"
         )
 
@@ -89,12 +92,12 @@ def test_get_table_identifiers_for_concept_scheme_error():
     path_to_cube = (
         _test_case_base_dir / "itis-industry-no-skos-inscheme.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     concept_scheme_url = "http://gss-data.org.uk/data/gss_data/trade/ons-international-trade-in-services#scheme/itis-industry"
 
     with pytest.raises(KeyError) as exception:
-        _ = code_list_inspector.get_table_identifiers_for_concept_scheme(
+        _ = code_list_repository.get_table_identifiers_for_concept_scheme(
             concept_scheme_url
         )
     assert (
@@ -112,11 +115,11 @@ def test_get_map_code_list_uri_to_label():
         / "code-list-with-uri-identifier-as-identifier"
         / "category.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     concept_scheme_uri = "category.csv#code-list"
 
-    result = code_list_inspector.get_map_code_list_uri_to_label(concept_scheme_uri)
+    result = code_list_repository.get_map_code_list_uri_to_label(concept_scheme_uri)
     assert result == {
         "boosting-productivity-pay-jobs-and-living-standards": "Boosting productivity, pay, jobs and living standards by growing the private sector",
         "spreading-opportunity-and-improving-public-services": "Spreading opportunity and improving public services",
@@ -134,12 +137,12 @@ def test_get_map_code_list_uri_to_label_duplicate_label():
         / "code-list-with-uri-identifier-as-identifier"
         / "category-duplicate-labels.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     concept_scheme_uri = "category-duplicate-labels.csv#code-list"
 
     with pytest.raises(ValueError) as exception:
-        _ = code_list_inspector.get_map_code_list_uri_to_label(concept_scheme_uri)
+        _ = code_list_repository.get_map_code_list_uri_to_label(concept_scheme_uri)
 
     assert (
         "Duplicate labels 'Restoring a sense of community, local pride and be…' in `Label` column for category-duplicate-labels.csv"
@@ -156,12 +159,12 @@ def test_get_map_code_list_uri_to_label_duplicate_uri():
         / "code-list-with-uri-identifier-as-identifier"
         / "category-duplicate-uris.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
+    code_list_repository = get_code_list_repository(path_to_cube)
 
     concept_scheme_uri = "category-duplicate-uris.csv#code-list"
 
     with pytest.raises(ValueError) as exception:
-        _ = code_list_inspector.get_map_code_list_uri_to_label(concept_scheme_uri)
+        _ = code_list_repository.get_map_code_list_uri_to_label(concept_scheme_uri)
 
     assert (
         "Duplicate URIs 'restoring-a-sense-of-community-local-pride-and-bel…' in `Uri Identifier` column for category-duplicate-uris.csv"
@@ -177,5 +180,5 @@ def test_get_primary_csv_url():
         / "pivoted-multi-measure-dataset"
         / "some-dimension.csv-metadata.json"
     )
-    code_list_inspector = get_code_list_inspector(path_to_cube)
-    assert code_list_inspector.get_primary_csv_url() == "some-dimension.csv"
+    code_list_repository = get_code_list_repository(path_to_cube)
+    assert code_list_repository.get_primary_csv_url() == "some-dimension.csv"
