+++ tests/unit/utils/test_csvdataset.py
@@ -16,7 +16,7 @@ from csvcubed.utils.csvdataset import (
     _melt_data_set,
     transform_dataset_to_canonical_shape,
 )
-from tests.helpers.inspectors_cache import (
+from tests.helpers.repository_cache import (
     get_csvw_rdf_manager,
     get_data_cube_repository,
 )
