+++ tests/unit/utils/test_tableschema.py
@@ -2,7 +2,7 @@ from csvcubedmodels.rdf.namespaces import CSVW
 from rdflib import Literal, URIRef
 
 from csvcubed.utils.tableschema import CsvWRdfManager
-from tests.helpers.inspectors_cache import get_csvw_rdf_manager
+from tests.helpers.repository_cache import get_csvw_rdf_manager
 from tests.unit.test_baseunit import get_test_cases_dir
 
 _test_case_base_dir = get_test_cases_dir() / "cli" / "inspect"
